xdg-ninja: change jq json source
There was an issue caused by inconsistencies on the implementation of
echo through different shells. Some, like bash, did not interpret \n
while others, like dash, did. This caused issue when sending the
information to jq to parse as it as JSON specifications forbids
multi-line strings.

A quick fix was applied by giving jq the path to the file to analyze
rather than grabbing it from stdin.