Add xargs to pass the list of programs to jq
Depending on the operating system the argument list has a limit. On
Windows this is 8191 characters:

https://learn.microsoft.com/en-us/troubleshoot/windows-client/shell-experience/command-line-string-limitation

The current list of programs in xdg-ninja already passes this limit,
so the script throws the following error:

> ./xdg-ninja.sh: line 228: /c/Users/USERNAME/.local/bin/jq:
> Argument list too long

This commit will use `find` and `xargs` to pass the list of programs
to `jq` to solve this issue. The list of program files needs to be
terminated by a null character (zero byte) so that `xargs` ignores the
whitespace in the "zen browser.json" filename. Both `find` and `xargs`
are used elsewhere in xdg-ninja already, so using them should not be a
problem.